tests/wpt/tests/dom/nodes/getElementsByClassName-09.htm

<!doctype html>
<html class="a A">
 <head>
  <title>document.getElementsByClassName(): case sensitive</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body class="a a">
  <div id="log"></div>
  <script>test(function() {
                 assert_array_equals(document.getElementsByClassName("A a"), [document.documentElement])
               })
  </script>
 </body>
</html>
